+++ tool/generate_changelog.dart
@@ -19,6 +19,8 @@ const _packageOption = 'package';
 /// To generate the CHANGELOG for WebDev:
 ///  `dart run generate_changelog.dart -p webdev`
 
+late String? accessToken;
+
 void main(List<String> arguments) async {
   final parser = ArgParser()
     ..addOption(
@@ -37,51 +39,68 @@ void main(List<String> arguments) async {
     return;
   }
 
-  generateChangelog(
-    package: package,
-  );
-  // if (exitCode != 0) {
-  //   _logWarning('Run terminated unexpectedly with exit code: $exitCode');
-  // }
-}
-
-void generateChangelog({
-  required String package,
-}) async {
-  print('Getting latest release for $package...');
-  final latestReleaseName = _latestReleaseName(package);
-  print('Looking up commit hash for $latestReleaseName...');
-  final commitHash = await _findCommitMatchingTagName(latestReleaseName);
-  final commits = await _getCommitsSince(commitHash);
-  final pulls = await _getPullsForPackage(commits, package: package);
-  for (final pull in pulls) {
-    print(pull.title);
+  accessToken = await _checkAccessToken();
+  if (accessToken == null) {
+    _logWarning(
+        'No access token found, will call Github APIs without authenticating.');
   }
 
+  _generateChangelog(
+    package: package,
+  );
+}
+
+Future<String?> _checkAccessToken() async {
+  final tokenFile = File('./gh_access_token.txt');
+  final tokenFileExists = await tokenFile.exists();
+  if (tokenFileExists) {
+    return tokenFile.readAsString();
+  }
+  return null;
+}
+
+void _generateChangelog({
+  required String package,
+}) async {
+  _logInfo('Getting latest release for $package...');
+  final latestReleaseName = _latestReleaseName(package);
+  _logInfo('Looking up commit for $latestReleaseName...');
+  final commit = await _findCommitMatchingTagName(latestReleaseName);
+  _logInfo('Getting all commits since ${commit.sha}...');
+  final commits = await _getCommitsSince(commit);
+  _logInfo('Getting the associated pulls for those commits...');
+  final pulls = await _getPullsForPackage(commits, package: package);
+  _logInfo('Writing pulls info to CHANGELOG...');
+  _writePullsToChangelog(pulls, package: package);
 }
 
 String _latestReleaseName(String package) {
   final changelog = File('../$package/CHANGELOG.md');
   final lines = changelog.readAsLinesSync();
-  // The third line contains the last release in the format "## XX.X.X":
+  // The third line contains the latest release in the format "## XX.X.X":
   return lines[3].trim().substring(3);
 }
 
-Future<String> _findCommitMatchingTagName(String tagMatcher) async {
-  final tags = (await _githubQuery('tags', params: {'per_page': '100'}) as List)
-      .cast<Map<String, dynamic>>();
+Future<_CommitInfo> _findCommitMatchingTagName(String tagMatcher) async {
+  final tags =
+      _listify(await _githubQuery('tags', params: {'per_page': '100'}));
   final matchingTag = tags.firstWhere((tag) {
     final tagName = tag['name'] as String;
     return tagName.contains(tagMatcher);
   });
-  return matchingTag['commit']['sha'] as String;
+  final sha = matchingTag['commit']['sha'] as String;
+  final commitInfo = await _githubQuery('commits/$sha');
+  final commitDate = commitInfo['commit']['committer']['date'] as String;
+  return _CommitInfo(dateStr: commitDate, sha: sha);
 }
 
-Future<List<String>> _getCommitsSince(String commitHash) async {
-  final commits =
-      (await _githubQuery('commits', params: {'since': commitHash}) as List)
-          .cast<Map<String, dynamic>>();
-  return commits.map((commit) => commit['sha'] as String).toList();
+Future<List<String>> _getCommitsSince(_CommitInfo commit) async {
+  final commits = _listify(
+      await _githubQuery('commits', params: {'since': commit.dateStr}));
+  return commits
+      .map((commit) => commit['sha'] as String)
+      .where((sha) => sha != commit.sha)
+      .toList();
 }
 
 Future<List<_PullInfo>> _getPullsForPackage(
@@ -90,7 +109,7 @@ Future<List<_PullInfo>> _getPullsForPackage(
 }) async {
   final pullsInfo = <_PullInfo>[];
   for (final commit in commits) {
-    final pulls = await _githubQuery('commits/$commit/pulls') as List;
+    final pulls = _listify(await _githubQuery('commits/$commit/pulls'));
     if (pulls.isEmpty) continue;
     final pull = pulls[0];
     final labels = _listify(pull['labels']);
@@ -100,7 +119,6 @@ Future<List<_PullInfo>> _getPullsForPackage(
       pullsInfo.add(
         _PullInfo(
           number: pull['number'] as int,
-          url: pull['url'] as String,
           title: pull['title'] as String,
         ),
       );
@@ -109,14 +127,39 @@ Future<List<_PullInfo>> _getPullsForPackage(
   return pullsInfo;
 }
 
+void _writePullsToChangelog(
+  List<_PullInfo> pulls, {
+  required String package,
+}) {
+  final pullInfoLines = pulls.map((pull) {
+    final formattedTitle = pull.title.trim().capitalize().addPeriod();
+    final pullNumber = pull.number;
+    final pullUrl = 'https://github.com/dart-lang/webdev/pull/$pullNumber';
+    return '- $formattedTitle - [#$pullNumber]($pullUrl)';
+  });
+  final changelog = File('../$package/CHANGELOG.md');
+  final changelogLines = changelog.readAsLinesSync();
+  final newContent = [
+    changelogLines[0], // The version header, e.g. "## XX.X.X"
+    '', // A new line
+    ...pullInfoLines,
+    '', // Another new line
+    ...changelogLines.sublist(3), // Previous versions
+  ].joinWithNewLine();
+  changelog.writeAsStringSync(newContent);
+}
+
 dynamic _githubQuery(String requestPath, {Map<String, dynamic>? params}) async {
   final uri = Uri.https(
     'api.github.com',
     '/repos/dart-lang/webdev/$requestPath',
     params,
   );
-  print('requesting $uri');
-  final githubResponse = await http.get(uri);
+  final githubResponse = accessToken == null
+      ? await http.get(uri)
+      : await http.get(uri, headers: {
+          HttpHeaders.authorizationHeader: 'token $accessToken',
+        });
   return jsonDecode(githubResponse.body);
 }
 
@@ -127,8 +170,14 @@ bool _hasLabelMatching(List<Map<String, dynamic>> labels,
   return matchingLabel != null;
 }
 
-List<Map<String, dynamic>> _listify(dynamic rawJson) =>
-    (rawJson as List).cast<Map<String, dynamic>>();
+List<Map<String, dynamic>> _listify(dynamic rawJson) {
+  try {
+    return (rawJson as List).cast<Map<String, dynamic>>();
+  } catch (error) {
+    _logWarning('Expected JSON list, instead got: $rawJson');
+    rethrow;
+  }
+}
 
 void _logInfo(String message) {
   stdout.writeln(message);
@@ -138,19 +187,44 @@ void _logWarning(String warning) {
   stderr.writeln(warning);
 }
 
+class _PullInfo {
+  int number;
+  String title;
+  _PullInfo({
+    required this.number,
+    required this.title,
+  });
+}
+
+class _CommitInfo {
+  String dateStr;
+  String sha;
+  _CommitInfo({
+    required this.dateStr,
+    required this.sha,
+  });
+}
+
+extension CapitalizeExtension on String {
+  String capitalize() {
+    if (isEmpty) return this;
+    final firstChar = this[0].toUpperCase();
+    final rest = substring(1);
+    return '$firstChar$rest';
+  }
+}
+
+extension PeriodExtension on String {
+  String addPeriod() {
+    if (isEmpty) return this;
+    final lastChar = this[length - 1];
+    if (lastChar == '.') return this;
+    return '$this.';
+  }
+}
+
 extension JoinExtension on List<String> {
   String joinWithNewLine() {
     return '${join('\n')}\n';
   }
-}
-
-class _PullInfo {
-  int number;
-  String url;
-  String title;
-  _PullInfo({
-    required this.number,
-    required this.url,
-    required this.title,
-  });
 }